Refactor: Convert single-file module to modular package structure
This commit refactors the Claude Notifications MCP Server from a single file
into a proper modular package structure. Key changes include:

- Create `notifications` package with submodules:
  - core/: Core functionality
  - platform/macos/: Platform-specific implementations
  - utils/: Utility functions
  - server.py: Main MCP server implementation
- Add proper entry points
- Update packaging configuration
- Update documentation
- Add tests for the modular structure
- Update GitHub Actions workflow for the new structure

Closes #7
