+++ __main__.py
@@ -1,0 +1,9 @@
+#!/usr/bin/env python3
+"""
+Main entry point for Claude Notifications MCP Server.
+"""
+
+from notifications.server import main
+
+if __name__ == "__main__":
+    exit(main())
