+++ notifications/__init__.py
@@ -1,0 +1,6 @@
+"""
+Claude Notifications MCP Server - Sound and visual notifications for Claude Desktop on macOS.
+"""
+
+# Version
+__version__ = "1.2.0"
