+++ notifications/core/__init__.py
@@ -1,0 +1,3 @@
+"""
+Core notification functionality for Claude Notifications MCP Server.
+"""
