+++ notifications/platform/__init__.py
@@ -1,0 +1,3 @@
+"""
+Platform-specific implementations for Claude Notifications MCP Server.
+"""
