+++ notifications/platform/macos/__init__.py
@@ -1,0 +1,3 @@
+"""
+macOS-specific implementations for Claude Notifications MCP Server.
+"""
