+++ notifications/utils/__init__.py
@@ -1,0 +1,3 @@
+"""
+Utility functions for Claude Notifications MCP Server.
+"""
